♻️ router선언부분 분리, app에 정의
