+++ src/main.tsx
@@ -1,24 +1,11 @@
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { PageA, PageB, PageC, NotFound } from '@pages';
+
 import App from './App';
 import './index.css';
 
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-    children: [
-      { index: true, element: <PageA /> },
-      { path: '*', element: <NotFound /> },
-      { path: '/pageb', element: <PageB /> },
-      { path: '/pagec', element: <PageC /> },
-    ],
-  },
-]);
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <RouterProvider router={router} />
+    <App />
   </React.StrictMode>
 );
